// src/permission.enum.ts
export const enum PermissionKey {
  // ViewCategoriesManangement = 'ViewCategoriesManangement',
  // ViewOwnCart = 'ViewOwnCart',
  ViewUsers = 'ViewUsers',
  AddUsers = 'AddUsers',
  UpdateUsers = 'UpdateUsers',
  DeleteUsers = 'DeleteUsers',
  Admin = 'Admin',
  Customer = 'Customer',
  SignUp = 'SignUp',
  LogIn = 'LogIn',
}